fixed bug in CAS parser where parse failed if only 1 reference contig.  Changed TraceFileServer writer URLs fixed tests.  fixed lots of compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@114 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/align/SmithWatermanAligner.java
@@ -167,7 +167,7 @@ public class SmithWatermanAligner implements Aligner<NucleotideGlyph>
          */
         alignment.setIdentity(identity / (double)alignmentLength);
         alignment.setScore(score.getScore(start));
-        alignment.setMatch(matchScore / (double)alignmentLength);
+        alignment.setMatch(matchScore / alignmentLength);
         
         /*
          * Update the alignment factory with the alignment stop points
